applicationUpdate

+++ Common/lib.swift
@@ -132,6 +132,8 @@ func copyDefaultsFiles() {
             let toUrl = dataUrl.appendingPathComponent(item)
             let toPath = dataUrl.appendingPathComponent(item).path
             
+            if applicationUpdate && !item.containsAny([".dct.",".xrefs."]) { continue }
+            
             try? FileManager.default.removeItem(at: toUrl)
             try? FileManager.default.copyItem(atPath: atPath, toPath: toPath)
         }
